+++ src/scr_python_console.cpp
@@ -15,7 +15,7 @@
 #include "shp_common.h"
 #include "shp_cut.h"
 #include "shp_fuse.h"
-#include "sketch.h"
+#include "skt.h"
 #include "utl_geom.h"
 
 #ifdef EZYCAD_HAVE_PYTHON
