+++ src/skt.h
