+++ src/skt_op_recorder.cpp
@@ -9,7 +9,7 @@
 #include "delta.h"
 #include "utl_geom.h"
 #include "gui_occt_view.h"
-#include "sketch.h"
+#include "skt.h"
 #include "skt_nodes.h"
 #include "skt_edge.h"
