+++ src/skt_topo.cpp
@@ -18,7 +18,7 @@
 #include <glm/glm.hpp>
 
 #include "gui_occt_view.h"
-#include "sketch.h"
+#include "skt.h"
 #include "skt_op_recorder.h"
 #include "skt_edge.h"
 #include "utl.h"
